add functionality to signal activities from definition

+++ src/tasks/ReceiveTask.js
@@ -29,12 +29,14 @@ export function ReceiveTaskBehaviour(activity) {
 
   function execute(executeMessage) {
     const content = executeMessage.content;
+    const {executionId} = content;
+
+    if (content.isRootScope) setupMessageHandling(executionId);
     if (loopCharacteristics && content.isRootScope) {
       return loopCharacteristics.execute(executeMessage);
     }
 
     let completed;
-    const {executionId} = content;
 
     const {message: referenceMessage, description} = resolveReference(executeMessage);
     broker.consume('message', onCatchMessage, {noAck: true, consumerTag: `_onmessage-${executionId}`});
@@ -50,22 +52,26 @@ export function ReceiveTaskBehaviour(activity) {
     function onCatchMessage(routingKey, message) {
       if (getPropertyValue(message, 'content.message.id') !== referenceMessage.id) return;
 
-      logger.debug(`<${executionId} (${id})> caught ${description}`);
-      broker.publish('event', 'activity.catch', cloneContent(content, {message: message.content.message}), {type: 'catch'});
+      const {type: messageType, correlationId} = message.properties;
+      broker.publish('event', 'activity.consumed', cloneContent(content, {message: {...message.content.message}}), {correlationId, type: messageType});
 
-      complete(message.content.message);
+      logger.debug(`<${executionId} (${id})> caught ${description}`);
+      broker.publish('event', 'activity.catch', cloneContent(content, {message: message.content.message}), {type: 'catch', correlationId});
+
+      complete(message.content.message, {correlationId});
     }
 
     function onApiMessage(routingKey, message) {
-      switch (message.properties.type) {
+      const {type: messageType, correlationId} = message.properties;
+      switch (messageType) {
         case 'message':
         case 'signal': {
-          return complete(message.content.message);
+          return complete(message.content.message, {correlationId});
         }
         case 'discard': {
           completed = true;
           stop();
-          return broker.publish('execution', 'execute.discard', cloneContent(content));
+          return broker.publish('execution', 'execute.discard', cloneContent(content), {correlationId});
         }
         case 'stop': {
           return stop();
@@ -73,16 +79,48 @@ export function ReceiveTaskBehaviour(activity) {
       }
     }
 
-    function complete(output) {
+    function complete(output, options) {
       completed = true;
       stop();
-      return broker.publish('execution', 'execute.completed', cloneContent(content, {output}));
+      return broker.publish('execution', 'execute.completed', cloneContent(content, {output}), options);
     }
 
     function stop() {
       broker.cancel(`_onmessage-${executionId}`);
       broker.cancel(`_api-${executionId}`);
-      broker.purgeQueue('message');
+    }
+  }
+
+  function setupMessageHandling(executionId) {
+    broker.subscribeTmp('api', '#.signal.*', onDelegateMessage, {noAck: true, consumerTag: `_api-delegated-${executionId}`}, {noAck: true});
+    broker.subscribeTmp('api', `activity.stop.${executionId}`, onStopApiMessage, {noAck: true, consumerTag: `_api-stop-${executionId}`, priority: 400});
+    broker.subscribeTmp('execution', 'execute.#', onComplete, {noAck: true, consumerTag: `_execution-complete-${executionId}`}, {noAck: true});
+
+    function onDelegateMessage(_, message) {
+      if (!message.properties.delegate) return;
+      broker.sendToQueue('message', message.content, message.properties);
+    }
+
+    function onStopApiMessage() {
+      stop(true);
+    }
+
+    function onComplete(routingKey, {content}) {
+      if (!content.isRootScope) return;
+      switch (routingKey) {
+        case 'execute.completed':
+        case 'execute.error':
+        case 'execute.discard':
+          stop();
+          break;
+      }
+    }
+
+    function stop(keepMessageQ) {
+      broker.cancel(`_api-delegated-${executionId}`);
+      broker.cancel(`_api-stop-${executionId}`);
+      broker.cancel(`_execution-complete-${executionId}`);
+      if (!keepMessageQ) broker.purgeQueue('message');
     }
   }
